Added level names and moved buttons back to the level editor script.
