+++ level-editor.py
@@ -5,7 +5,7 @@ import configparser
 
 from Utils import *
 from GameState import GameState, LevelLoader
-from Display import *
+from LevelDisplay import *
 
 TOOLBAR_THICKNESS = 0.1
 TILEBAR_THICKNESS = 0.2
@@ -32,7 +32,7 @@ states = [GameState()]
 
 screen = pygame.display.set_mode(screen_size)
 display_rect = [0, TOOLBAR_THICKNESS*screen_height, screen_width - TILEBAR_THICKNESS*screen_width, screen_height - 2*TOOLBAR_THICKNESS*screen_height]
-display = Display(screen, display_rect)
+display = LevelDisplay(screen, display_rect)
 display.debug_grid = True
 
 class ButtonType(Enum):
@@ -80,6 +80,62 @@ button_type_to_tile_type_map = {
     ButtonType.TILE_BLOCK_E_PIT: 'e',
     ButtonType.TILE_BLOCK_F_PIT: 'f'
 }   
+
+class Button:
+
+    def __init__(self, button_type, screen, x, y, width, height, image_filename, action_func=None):
+        self.button_type = button_type
+        self.screen = screen
+        self.x = x
+        self.y = y
+        self.width = width
+        self.height = height
+        self.image_filename = image_filename
+        self.image = pygame.image.load(image_filename)
+        self.action_func = action_func
+        self.border_colour = GREY
+
+    def position_inside_button(self, x, y):
+        return self.x < x < self.x + self.width and self.y < y < self.y + self.height
+        
+    def draw(self):
+        # Draw the main button border
+        border_offset = int((BUTTON_BORDER_THICKNESS * self.width)/2)
+        top_left = (self.x + border_offset, self.y + border_offset)
+        top_right = (self.x + (self.width - border_offset), self.y + border_offset)
+        bottom_left = (self.x + border_offset, self.y + (self.height - border_offset))
+        bottom_right = (self.x + (self.width - border_offset), self.y + (self.height - border_offset))
+        pygame.draw.line(self.screen, self.border_colour, top_left, top_right, border_offset*2)
+        pygame.draw.line(self.screen, self.border_colour, top_left, bottom_left, border_offset*2)
+        pygame.draw.line(self.screen, self.border_colour, bottom_left, bottom_right, border_offset*2)
+        pygame.draw.line(self.screen, self.border_colour, top_right, bottom_right, border_offset*2)
+        # Draw the image on the button
+        scaled = pygame.transform.scale(self.image, (self.width - border_offset*6, self.height - border_offset*6))
+        self.screen.blit(scaled, [self.x + border_offset*3, self.y + border_offset*3, self.width - border_offset*2, self.height - border_offset*2])
+
+class ToggleButtonGroup:
+
+    def __init__(self, buttons):
+        self.buttons = buttons
+        self.active = 0
+        self.update_buttons()
+    
+    def update_buttons(self):
+        for i, button in enumerate(self.buttons):
+            if i == self.active:
+                button.border_colour = LIGHTER_GREY
+            else:
+                button.border_colour = GREY
+
+    def check_for_new_active(self, mouse_pos):
+        for i, button in enumerate(self.buttons):
+            if button.position_inside_button(mouse_pos[0], mouse_pos[1]):
+                self.active = i
+                self.update_buttons()
+                break
+
+    def get_active_button(self):
+        return self.buttons[self.active]
 
 # Set up regular action buttons
 action_buttons = list()
